Couple of working tests
Please see baseBackEnd_Functional_Tests.java for examples of tests.
Please expand on this!
Currently our testing scheme is using Check.java which is a copy of the
functions in check.jsp. So for now any time functions in check.jsp are
changed, we'll need to copy the source code to Check.java if we want to
run unit tests.